+++ packages/jaspr_cli/lib/src/dev/dev_proxy.dart
@@ -55,16 +55,33 @@ class DevProxy {
   }
 
   void _listenToBuildResults() async {
-    if (reload == ReloadConfiguration.hotReload) {
-      await for (final buildResult in buildResults) {
-        if (buildResult.status == BuildStatus.succeeded) {
-          for (final clientConnection in _clientConnections.values) {
+    await for (final buildResult in buildResults) {
+      if (buildResult.status == BuildStatus.succeeded) {
+        for (final clientConnection in _clientConnections.values) {
+          if (reload == ReloadConfiguration.hotReload) {
             await clientConnection.performHotReload();
           }
+        }
+
+        for (final callback in _postReloadCallbacks) {
+          // ignore: avoid_dynamic_calls
+          callback(buildResult);
         }
       }
     }
   }
+
+  final List<void Function(BuildResult result)> _postReloadCallbacks = [];
+
+  void registerPostReloadCallback(void Function(BuildResult result) callback) {
+    _postReloadCallbacks.add(callback);
+  }
+
+  void unregisterPostReloadCallback(void Function(BuildResult result) callback) {
+    _postReloadCallbacks.remove(callback);
+  }
+
+  List<ClientConnection> getClientConnections() => _clientConnections.values.toList();
 
   ClientConnection? getClientConnection(String? appId) {
     if (appId == null) return null;
@@ -99,18 +116,18 @@ class DevProxy {
           }
         });
       }
-      final result = results.results.firstWhere((result) => result.target == target);
-      switch (result.status) {
-        case daemon.BuildStatus.started:
-          return BuildResult(status: BuildStatus.started);
-        case daemon.BuildStatus.failed:
-          return BuildResult(status: BuildStatus.failed);
-        case daemon.BuildStatus.succeeded:
-          return BuildResult(status: BuildStatus.succeeded);
-        default:
-          break;
+      final resultForTarget = results.results.where((result) => result.target == target).firstOrNull;
+      final result = switch (resultForTarget?.status) {
+        daemon.BuildStatus.started => BuildResult(status: BuildStatus.started),
+        daemon.BuildStatus.failed => BuildResult(status: BuildStatus.failed),
+        daemon.BuildStatus.succeeded => BuildResult(status: BuildStatus.succeeded),
+        _ => null,
+      };
+      if (result == null) {
+        throw StateError('Unexpected Daemon build result: $result');
       }
-      throw StateError('Unexpected Daemon build result: $result');
+      _changedAssetsForBuildResult[result] = results.changedAssets?.toList();
+      return result;
     });
 
     var cascade = Cascade();
@@ -439,4 +456,10 @@ class ClientConnection {
     _vmServiceSub?.cancel();
     _debugConnection?.close();
   }
+}
+
+final Expando<List<Uri>> _changedAssetsForBuildResult = Expando<List<Uri>>();
+
+extension BuildResultExt on BuildResult {
+  List<Uri>? get changedAssets => _changedAssetsForBuildResult[this];
 }